Implement order management mudole and place order feature

+++ BookStoreWebsite/src/com/bookstore/controller/frontend/shoppingcart/ShoppingCart.java
@@ -38,8 +38,8 @@ public class ShoppingCart {
 		return total;
 	}
 
-	public double getTotalAmount() {
-		double total = 0;
+	public float getTotalAmount() {
+		float total = 0;
 
 		Iterator<Book> iterator = cart.keySet().iterator();
